Program was not compiling. An attempt to find and fix the bug causing this was made on 22/11/2020s
Two projects were merged. Necessary code could not be accessed. This contributed to unfunctional code.s

I removed unsued event handlers and methods#

Sql connection was made every time an action was performed. This made the program significantly slow.

Application was still running in the background after closing

Added data to the Users table in the database but the username field  did not exist causing the program to crash internally.

Added the username column to the table